requirements.txt
fastapi
uvicorn
python-pptx
lightgbm
sentence-transformers
textstat
scikit-learn
numpy
pandas
joblib
